remove introspectUtils, move its functions to KspIntrospectionContext, make `_nodes` private in BaseIntrospectionContext
